pyproject.toml
[project]
name = "nicegui-blog"
version = "0.1.0"
description = "A reusable, file-based blog template with NiceGUI"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "nicegui>=2.22.1",
    "markdown>=3.7.0",
    "python-frontmatter>=1.1.0",
    "pygments>=2.18.0",
    "pymdown-extensions>=10.13.0",
    "pyinstaller>=6.14.2",
]
requires-python = "~=3.13"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "RUF", # ruff-specific rules
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "RUF012", # mutable class attributes
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["app"]